add all the files of spark quiz

+++ app/Http/Controllers/Web/Admin/Property/PropertyTypeController.php
@@ -2,17 +2,14 @@
 
 namespace App\Http\Controllers\Web\Admin\Property;
 
-use App\DataTables\PropertyTypesDataTable;
-use App\Http\Controllers\API\V1\Admin\AdminAPIBaseController;
-use App\Http\Controllers\Controller;
+use App\DataTables\Property\PropertyTypesDataTable;
+use App\Http\Controllers\Web\Admin\AdminBaseController;
 use App\Http\Requests\API\V1\Admin\Property\StorePropertyTypeRequest;
-use App\Http\Resources\API\V1\Admin\Property\PropertyTypeCollection;
+use App\Http\Requests\API\V1\Admin\Property\UpdatePropertyTypeRequest;
 use App\Http\Resources\API\V1\Admin\Property\PropertyTypeResource;
 use App\Models\V1\Property\PropertyType;
-use Illuminate\Http\Request;
-use Yajra\DataTables\Facades\DataTables;
 
-class PropertyTypeController extends Controller
+class PropertyTypeController extends AdminBaseController
 {
 
     public function index(PropertyTypesDataTable $dataTable)
@@ -38,75 +35,88 @@ class PropertyTypeController extends Controller
         $propertyType->fill($request->all());
         $propertyType->created_by_id = $created_by_id;
         $propertyType->save();
-        //return response()->json($propertyType);
-        //return new PropertyTypeResource($propertyType);
+
+
+        $displayUrl = route('web.admin.property-types.show', $propertyType->id);
+        $this->flashStoredMsg($displayUrl);
+        return redirect()->back()->with(['rID' =>$propertyType->id ]);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int $id
+     * @param int $id
      * @return PropertyTypeResource
      */
     public function show($id)
     {
-        $propertyType = PropertyType::with('properties.events.eventType')->findOrFail($id);
-        return new PropertyTypeResource($propertyType);
-        //$propertyType=PropertyType::with('properties.owner')->findOrFail($id);
-        //return new PropertyTypeResource($propertyType);
+        $propertyType = PropertyType::with('properties')->findOrFail($id);
+        $this->authorize('show',$propertyType);
+        return view('admin.property-types.show', compact('propertyType'));
+    }
+
+    public function edit($id)
+    {
+        $propertyType = PropertyType::findOrFail($id);
+        return view('admin.property-types.edit', compact('propertyType'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param StorePropertyTypeRequest $request
-     * @param  int $id
+     * @param UpdatePropertyTypeRequest $request
+     * @param int $id
      * @return PropertyTypeResource
      */
-    public function update(StorePropertyTypeRequest $request, $id)
+    public function update(UpdatePropertyTypeRequest $request, $id)
     {
         $propertyType = PropertyType::findOrFail($id);
         $propertyType->fill($request->all());
+        $this->authorize('update',$propertyType);
         $propertyType->save();
-        return new PropertyTypeResource($propertyType);
+        $displayUrl = route('web.admin.property-types.show', $propertyType->id);
+        $this->flashUpdatedMsg($displayUrl);
+        return redirect()->back()->with(['rID' => $propertyType->id]);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
-    //PERMANENT DELETE
+
     public function destroy($id)
     {
-        //$propertyType=PropertyType::onlyTrashed()->where('id',$id)->forceDelete();
-        $propertyType = PropertyType::findOrFail($id);
+        $propertyType = PropertyType::withTrashed()->findOrFail($id);
+        // $this->authorize('destroy',$propertyType);
         $propertyType->forceDelete();
         return response('PERMANENTLY DELETED');
     }
 
-    //soft delete
+    /**
+     * @param $id
+     * @return \Illuminate\Http\Response
+     */
     public function trash($id)
     {
         $propertyType = PropertyType::find($id);
+        $this->authorize('trash',$propertyType);
         $propertyType->delete();
         return response()->noContent();
     }
 
-    //restore data
+    /**
+     * @param $id
+     * @return \Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
+     */
     public function restore($id)
     {
-        $propertyType = PropertyType::withTrashed()->where('id', $id)->restore();
+        //$propertyType = PropertyType::withTrashed()->where('id', $id)->restore();
+        $propertyType = PropertyType::withTrashed()->findOrFail($id);
+        $this->authorize('restore',$propertyType);
+        $propertyType->restore();
         return response('SUCCESSFULLY RESTORED');
     }
 
-
-    //searching anything
-    public function search(Request $request)
-    {
-        $propertyTypeName = $request->get('name');
-        $propertyTypes = PropertyType::where('name', 'LIKE', '%' . $propertyTypeName . '%')->orderBy('id')->paginate(2);
-        return response($propertyTypes);
-    }
 }
